requirements.txt
rpy2==3.5.1
Django==4.0.4
django-extensions==3.1.5
django-ninja==0.17.0
